Completely reorganize title page layout
This restyles the navigation menu and moves the server info side panel to
between the navigation menu and the title page main content.

Where the info panel belongs is up for debate. The current placement takes
up considerable valuable screen real-estate and the reality is that the
information may not be very valuable. Regular players won't benefit from
it, but then, they won't benefit much from the main site at all except as
a link hub. It has value to new players that have decided to try the
server, but to new players who have not yet made that decision the
information gets in the way of the presentational content.