Replace inline-blocks with flexbox layout
Over time, flexbox is simpler than inline-block, which in turn is
simpler than float. The precise meaning of this is that flexbox *tends*
to require less code and lends more easily to responsive design, and
further, is not encumbered by the annoying whitespace sensitivity that
plagues inline-block. The change is not necessary and has a cost, but is
basically an investment in the future as flexbox support will only
improve.

Browser support for flexbox isn't *great* but it is *good* [1].
Currently, Safari, iOS Safari, and IE10 require special handling.

     * Safari issues appear solvable only with vendor prefixes, so an
      exception has been made to fix this; generally, vendor prefixes
      are disallowed.

     * IE10's issues can't be solved purely with vendor prefixing, but
      any machine that supports IE10 also supports IE11, which has full
      flexbox support, so IE10 is not getting any special treatment.

At some unspecified future date, an automated build step -- probably
Gulp -- may be added. If so, Autoprefixer [2] will be included, at which
point Safari's prefixes must be removed. That should also solve the IE10
issues, although those have very low priority given the above.

[1] http://caniuse.com/#search=flex
[2] https://github.com/postcss/autoprefixer