Pad all text content with 1em
Especially when the status element is not displayed, text content gets
too close to borders.

h1.margin-top alone won't work since it creates a gap in the background
colour. With padding-top it's fine but also redundant. Between .content
and h1, the padding placement is largely irrelevant, but the bottom
padding that comes with this solution is nice.