Fix padding inside flexbox overflowing in IE
The navigation bar's items display overflow in landscape mode in IE up
to and including IE 11. This commit provides a work-around by forcing
appropriate min-widths on all items. This is generally what flexbox was
intended to avoid in the first place, but as long as more than one item
can share a min-width, flexbox is still a more flexible solution than
hardcoded widths, and it will respond to layout changes much better.

In IE < 12, elements with padding, when nested inside a flex-item, can
cause overflow. Flexbug 7 [1] demonstrates a related but slightly
different problem, and none of its work-arounds solve this particular
case.

[1] https://github.com/philipwalton/flexbugs/blob/f642b54ceb1079cb98cd0426c3ef7401989286de/README.md#7-flex-basis-doesnt-account-for-box-sizingborder-box