Refactor page navigation item lines
In preparation for site generation, refactor page the navigation items
so the file path is on the same line as the CSS classes, to facilitate
stream processing, and so the current-item CSS class consistently comes
immediately after the navigation--item marker. The latter is to simplify
the string manipulation necessary.

Off-site navigation items don't need any processing and can stay as they
are, which makes them more robust to change.